+++ CHANGELOG.md
@@ -5,8 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
-
+- manipulator: allows to bump versions and set prerelease and build metadata
 ### Changed
+- cli is not located in the cmd folder
 
 ## [0.1.0] 2016-12-30
 ### Added